Added a monster class

+++ include/Stats.h
@@ -20,12 +20,12 @@ namespace monsterbattle
     {
         struct Stats
         {
-            uint16_t health;    //Health stat
-            uint16_t defense;   //Defense stat
-            uint16_t attack;    //Attack stat
+            uint8_t health;    //Health stat
+            uint8_t defense;   //Defense stat
+            uint8_t attack;    //Attack stat
             float precision;    //Precision, a normalized value showing the accuracy
             float avoidance;    //Avoidancem a normalized value displaying the probability of avoiding an attack
-            uint16_t speed;     //Speed stat
+            uint8_t speed;     //Speed stat
         };
     }  
 }
